Deploy keepalived only when ceph-nfs is deployed
This is a WIP solution to provide stable VIPs to the ganesha instances
using keepalived but not haproxy.
This solution relies on a new config option that should be set at
cluster lever and allows to skip the haproxy deployment for ceph-nfs
but not for RGW.
The current code provides unit test to make sure we are not breaking
the rgw deployment.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>